Conflicts:
	.gitignore
	thirdParty/plcrashreporter-1.2-rc4/Tools/google-toolbox-for-mac-trunk-r582/SpotlightPlugins/AppleScript/TestData/test.scpt.tar

+++ kidozen.client/kidozen.client/NSString+Utilities.m
@@ -15,7 +15,7 @@
 - (int) indexOf:(NSString *)text {
     NSRange range = [self rangeOfString:text];
     if ( range.length > 0 ) {
-        return range.location;
+        return (int)range.location;
     } else {
         return -1;
     }
@@ -149,7 +149,7 @@
     unsigned char md5Buffer[CC_MD5_DIGEST_LENGTH];
     
     // Create 16 byte MD5 hash value, store in buffer
-    CC_MD5(ptr, strlen(ptr), md5Buffer);
+    CC_MD5(ptr, (unsigned int)strlen(ptr), md5Buffer);
     
     // Convert MD5 value in the buffer to NSString of hex values
     NSMutableString *output = [NSMutableString stringWithCapacity:CC_MD5_DIGEST_LENGTH * 2];
